+++ src/common/guards/jwt-guards/index.ts
@@ -1,0 +1,1 @@
+export * from './def-jwt-guard';
