+++ src/modules/vision/vision.controller.ts
@@ -1,9 +1,9 @@
 import { Body, Controller, Post, UseFilters, UseGuards } from '@nestjs/common';
 
-import { PortGuard } from '@common/guards/request-port-guard/request-port.guard';
-import { HttpExceptionFilter } from '@common/exception/httpException.filter';
-import { errorHandler } from '@common/helpers/error-handler.helper';
 import { OnPort } from '@common/decorators/port/port.decorator';
+import { PortGuard } from '@common/guards/request-port-guard';
+import { HttpExceptionFilter } from '@common/exception';
+import { errorHandler } from '@common/helpers';
 import { VISION_CONTROLLER, VISION_ROUTES } from '@libs/contracts/api/controllers/vision';
 import { XRAY_INTERNAL_API_PORT } from '@libs/contracts/constants';
